backend/requirements.txt
fastapi
 uvicorn[standard]
 pydantic
 python-multipart
 pypdf
 python-docx
 beautifulsoup4
 requests